+++ process_bigraph/core/tests.py
@@ -6,7 +6,7 @@ import random
 
 from process_bigraph.core.composite import Process, Step, Composite
 from process_bigraph.core.composite import merge_collections
-from process_bigraph.core.type_system import types
+from process_bigraph.type_system.utils import types
 
 
 class IncreaseProcess(Process):
