+++ process_bigraph/type_system/data_model.py
@@ -1,151 +1,5 @@
-"""
-=============
-Process Types
-=============
-"""
-
 from bigraph_schema import TypeSystem, get_path, establish_path, set_path
-from process_bigraph.core.registry import protocol_registry
-
-
-
-process_interval_schema = {
-    '_type': 'float',
-    '_apply': 'set',
-    '_default': '1.0'}
-
-
-# TODO: implement these
-def apply_process(current, update, bindings=None, types=None):
-    process_schema = dict(types.access('process'))
-    process_schema.pop('_apply')
-    return types.apply(
-        process_schema,
-        current,
-        update)
-
-
-def divide_process(value, bindings=None, types=None):
-    return value
-
-
-def serialize_process(value, bindings=None, types=None):
-    # TODO -- need to get back the protocol: address and the config
-    return value
-
-
-DEFAULT_INTERVAL = 1.0
-
-
-TYPE_SCHEMAS = {
-    'float': 'float'}
-
-
-def deserialize_process(serialized, bindings=None, types=None):
-    deserialized = serialized.copy()
-    protocol, address = serialized['address'].split(':', 1)
-
-    if 'instance' in deserialized:
-        instantiate = type(deserialized['instance'])
-    else:
-        process_lookup = protocol_registry.access(protocol)
-        if not process_lookup:
-            raise Exception(f'protocol "{protocol}" not implemented')
-
-        instantiate = process_lookup(address)
-        if not instantiate:
-            raise Exception(f'process "{address}" not found')
-
-    config = types.hydrate_state(
-        instantiate.config_schema,
-        serialized.get('config', {}))
-
-    interval = types.deserialize(
-        process_interval_schema,
-        serialized.get('interval'))
-
-    if not 'instance' in deserialized:
-        process = instantiate(config)
-        deserialized['instance'] = process
-
-    deserialized['config'] = config
-    deserialized['interval'] = interval
-
-    return deserialized
-
-
-def deserialize_step(serialized, bindings=None, types=None):
-    deserialized = serialized.copy()
-    protocol, address = serialized['address'].split(':', 1)
-
-    if 'instance' in deserialized:
-        instantiate = type(deserialized['instance'])
-    else:
-        process_lookup = protocol_registry.access(protocol)
-        if not process_lookup:
-            raise Exception(f'protocol "{protocol}" not implemented')
-
-        instantiate = process_lookup(address)
-        if not instantiate:
-            raise Exception(f'process "{address}" not found')
-
-    config = types.hydrate_state(
-        instantiate.config_schema,
-        serialized.get('config', {}))
-
-    if not 'instance' in deserialized:
-        process = instantiate(config)
-        deserialized['instance'] = process
-
-    deserialized['config'] = config
-
-    return deserialized
-
-
-process_types = {
-    'protocol': {
-        '_super': 'string'},
-
-    # TODO: step wires are directional ie, we make a distinction
-    #   between inputs and outputs, and the same wire cannot be both
-    #   an input and output at the same time
-    'step': {
-        '_super': ['edge'],
-        '_apply': 'apply_process',
-        '_serialize': 'serialize_process',
-        '_deserialize': 'deserialize_step',
-        '_divide': 'divide_process',
-        '_description': '',
-        # TODO: support reference to type parameters from other states
-        'address': 'protocol',
-        'config': 'tree[any]'},
-
-    'process': {
-        '_super': ['edge'],
-        '_apply': 'apply_process',
-        '_serialize': 'serialize_process',
-        '_deserialize': 'deserialize_process',
-        '_divide': 'divide_process',
-        '_description': '',
-        # TODO: support reference to type parameters from other states
-        'interval': process_interval_schema,
-        'address': 'protocol',
-        'config': 'tree[any]'},
-}
-
-
-def register_process_types(types):
-    types.apply_registry.register('apply_process', apply_process)
-    types.serialize_registry.register('serialize_process', serialize_process)
-    types.deserialize_registry.register('deserialize_process', deserialize_process)
-    types.divide_registry.register('divide_process', divide_process)
-
-    types.deserialize_registry.register('deserialize_step', deserialize_step)
-
-    for process_key, process_type in process_types.items():
-        types.type_registry.register(process_key, process_type)
-
-    return types
+from process_bigraph.type_system.utils import register_process_types, TYPE_SCHEMAS
 
 
 class ProcessTypes(TypeSystem):
@@ -153,14 +7,11 @@ class ProcessTypes(TypeSystem):
         super().__init__()
         register_process_types(self)
 
-
     # def serialize(self, schema, state):
     #     return ''
 
-
     # def deserialize(self, schema, encoded):
     #     return {}
-
 
     def infer_wires(self, ports, state, wires, top_schema=None, path=None):
         top_schema = top_schema or {}
@@ -168,7 +19,8 @@ class ProcessTypes(TypeSystem):
 
         for port_key, port_wires in wires.items():
             if isinstance(ports, str):
-                import ipdb; ipdb.set_trace()
+                import ipdb;
+                ipdb.set_trace()
             port_schema = ports.get(port_key, {})
             # port_wires = wires.get(port_key, ())
             if isinstance(port_wires, dict):
@@ -200,7 +52,6 @@ class ProcessTypes(TypeSystem):
                     destination[destination_key] = port_schema
 
         return top_schema
-
 
     def infer_schema(self, schema, state, top_state=None, path=None):
         '''
@@ -278,7 +129,8 @@ class ProcessTypes(TypeSystem):
             else:
                 for key, value in state.items():
                     inner_path = path + (key,)
-                    if get_path(schema, inner_path) is None or get_path(state, inner_path) is None or (isinstance(value, dict) and '_type' in value):
+                    if get_path(schema, inner_path) is None or get_path(state, inner_path) is None or (
+                            isinstance(value, dict) and '_type' in value):
                         schema, top_state = self.infer_schema(
                             schema,
                             value,
@@ -306,14 +158,14 @@ class ProcessTypes(TypeSystem):
                 destination[path_key] = type_schema
 
         return schema, top_state
-        
 
     def infer_edge(self, schema, wires):
         schema = schema or {}
         edge = {}
 
         if isinstance(wires, str):
-            import ipdb; ipdb.set_trace()
+            import ipdb;
+            ipdb.set_trace()
 
         for port_key, wire in wires.items():
             if isinstance(wire, dict):
@@ -326,7 +178,6 @@ class ProcessTypes(TypeSystem):
 
         return edge
 
-
     def initialize_edge_state(self, schema, path, edge):
         initial_state = edge['instance'].initial_state()
         ports = get_path(schema, path + ('_ports',))
@@ -336,7 +187,6 @@ class ProcessTypes(TypeSystem):
             edge['wires'],
             path[:-1],
             initial_state)
-        
 
     def hydrate_state(self, schema, state):
         if isinstance(state, str) or '_deserialize' in schema:
@@ -362,16 +212,13 @@ class ProcessTypes(TypeSystem):
 
         return result
 
-
     def hydrate(self, schema, state):
         # TODO: support partial hydration (!)
         hydrated = self.hydrate_state(schema, state)
         return self.fill(schema, hydrated)
 
-
     def dehydrate(self, schema):
         return {}
-
 
     def lookup_address(self, address):
         protocol, config = address.split(':')
